Address feedback from PR request to upstream Minor other fixes
