+++ Assets/Scripts/Prototype/Parts/PartAttributes.cs
@@ -1,4 +1,4 @@
-namespace Arkship.Parts
+namespace Kosmos.Prototype.Parts
 {
     //TODO - This should probably have a load of sub-classes for different types of files
     // with their own parameters (e.g. [FloatTweakable(min, max)])
