+++ Assets/Scripts/Prototype/Parts/PartDictionary.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Reflection;
 using UnityEngine;
 
-namespace Arkship.Parts
+namespace Kosmos.Prototype.Parts
 {
     public class PartTypeReflectionEntry
     {
