+++ Assets/Scripts/Prototype/Parts/PartSocket.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Arkship.Parts
+namespace Kosmos.Prototype.Parts
 {
     public class PartSocket : MonoBehaviour
     {
