+++ Assets/Scripts/Prototype/Parts/RocketPart.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.Serialization;
 
-namespace Arkship.Parts
+namespace Kosmos.Prototype.Parts
 {
     public class RocketPart : PartBase
     {
