+++ Assets/StreamingAssets/Vehicles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b167e39d408af734ea004eb7064aad02
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
