+++ Assets/StreamingAssets/Vehicles/F1.veh
@@ -1,0 +1,47 @@
+{
+    "Parts": [
+        {
+            "PartDefGuid": "947499b0492844be859aef3c4ad3502a",
+            "LocalPosition": {
+                "x": 0.0,
+                "y": 0.0,
+                "z": 0.0
+            },
+            "LocalRotation": {
+                "x": 0.0,
+                "y": 0.0,
+                "z": 0.0,
+                "w": 1.0
+            },
+            "Tweakables": []
+        },
+        {
+            "PartDefGuid": "6fa35f53ba9b4d78b0741eaa5e7585a7",
+            "LocalPosition": {
+                "x": 0.0,
+                "y": -1.502000093460083,
+                "z": 0.0
+            },
+            "LocalRotation": {
+                "x": 0.0,
+                "y": 0.0,
+                "z": 0.0,
+                "w": 1.0
+            },
+            "Tweakables": [
+                {
+                    "Name": "_currentFuel",
+                    "Value": "1"
+                }
+            ]
+        }
+    ],
+    "Connections": [
+        {
+            "ParentPartIndex": 0,
+            "ParentSocketIndex": 0,
+            "ChildPartIndex": 1,
+            "ChildSockedIndex": 0
+        }
+    ]
+}
