+++ Assets/StreamingAssets/Vehicles/F1.veh.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5c4e0628ce9531c4a82c763ec2232e58
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
